Move open_webui to hayhooks root folder
